+++ libraries/date_time.rb
@@ -12,39 +12,43 @@
 require_relative 'base_resource'
 
 module IloCookbook
-  # Class for Ilo Date Time Actions
+  # Class for iLO Date Time Actions
   class DateTime < BaseResource
     resource_name :ilo_date_time
 
     load_base_properties
     property :time_zone, String
     property :ntp_servers, Array
-    property :use_ntp, [TrueClass, FalseClass]
+    property :use_dhcpv4, [TrueClass, FalseClass]
 
     action :set do
-      raise 'Please provide the :time_zone, :ntp_servers and/or :use_ntp properties!' unless time_zone || ntp_servers || !use_ntp.nil?
+      error_msg = 'Please provide the :time_zone and/or :ntp_servers properties or the :use_dhcpv4 property!'
+      raise error_msg unless time_zone || ntp_servers || !use_dhcpv4.nil?
       load_sdk
       ilos.each do |ilo|
         client = build_client(ilo)
-        if time_zone
-          cur_val = client.get_time_zone
-          next if cur_val == time_zone
-          converge_by "Set ilo #{client.host} time zone from '#{cur_val}' to '#{time_zone}'" do
-            client.set_time_zone(time_zone)
-          end
-        end
-        unless use_ntp.nil?
+        unless use_dhcpv4.nil?
           cur_val = client.get_ntp
-          next if cur_val == use_ntp
-          converge_by "Set ilo #{client.host} NTP use from '#{cur_val}' to '#{use_ntp}'" do
-            client.set_ntp(use_ntp)
+          unless cur_val == use_dhcpv4
+            converge_by "#{use_dhcpv4 ? 'Enable' : 'Disable'} iLO #{client.host} DHCPv4-supplied time settings" do
+              client.set_ntp(use_dhcpv4)
+            end
           end
         end
         if ntp_servers
           cur_val = client.get_ntp_servers
-          next if cur_val == ntp_servers
-          converge_by "Set ilo #{client.host} NTP Servers from '#{cur_val}' to '#{ntp_servers}'" do
-            client.set_ntp_servers(ntp_servers)
+          unless cur_val == ntp_servers
+            converge_by "Set iLO #{client.host} NTP servers from '#{cur_val}' to '#{ntp_servers}'" do
+              client.set_ntp_servers(ntp_servers)
+            end
+          end
+        end
+        if time_zone
+          cur_val = client.get_time_zone
+          unless cur_val == time_zone
+            converge_by "Set iLO #{client.host} time zone from '#{cur_val}' to '#{time_zone}'" do
+              client.set_time_zone(time_zone)
+            end
           end
         end
       end